Module cpu::context

x86_64 execution contexts.

This is inteded to be general-purpose and composable, so that the same code can be reused for interrupts and for multithreading.

Structs

Context
Thread execution context

InterruptFrame

Registers
Registers pushed to the stack when handling an interrupt or context switch.