Module cpu::control_regs

x86 and x86_64 control registers

Modules

cr0
%cr0 contains flags that modify basic processor operation.

cr2
$cr2 contains the page fault linear address

cr3
%cr3 contains the page table root pointer

cr4
%cr4 contains flags that control protected mode execution.

Structs

CrState
A struct bundling together a snapshot of the control registers state.

Functions

dump
Dump the current contents of the control registers to a CrState.