Module cpu::control_regs::cr4

%cr4 contains flags that control protected mode execution.

Structs

Flags
Bitflags present in $cr4

Constants

DE
Debugging Extensions

FSGSBASE
Enables the instructions RDFSBASE, RDGSBASE, WRFSBASE, and WRGSBASE.

MCE
Machine Check Exception

OSFXSR
Operating system support for FXSAVE and FXRSTOR instructions

OSXMMEXCPT
Operating System Support for Unmasked SIMD Floating-Point Exceptions

OSXSAVE
XSAVE and Processor Extended States Enable

PAE
Physical Address Extension

PCE
Performance-Monitoring Counter enable

PCIDE
PCID Enable

PGE
Page Global Enabled

PKE
Protection Key Enable

PSE
Page Size Extension

PVI
Protected-mode Virtual Interrupts

SMAP
Supervisor Mode Access Protection Enable

SMEP
Supervisor Mode Execution Protection Enable

SMXE
Safer Mode Extensions Enable

TSD
Time Stamp Disable

VME
Virtual 8086 Mode Extensions

VMXE
Virtual Machine Extensions Enable

Functions

disable_timestamp⚠
If disabled, the RTDSC instruction can only be executed in Ring 0.

is_timestamp_disabled⚠
If disabled, the RTDSC instruction can only be executed in Ring 0.

read⚠
Read the current value from $cr4.

write⚠
Write a value to $cr4.