Trait cpu::dtable::DTable

pub trait DTable: Sized {
    type Entry: Sized;
    fn entry_count(&self) -> usize;
    fn load(&'static self);

    fn get_ptr(&self) -> Pointer<Self> { ... }
}

A descriptor table (IDT or GDT).

The IA32 architecture uses two descriptor table structures, the GDT (Global Descriptor Table), which is used for configuring segmentation, and the IDT (Interrupt Descriptor Table), which tells the CPU where interrupt service routines are located.

As SOS relies on paging rather than segmentation for memory protection on both 32-bit and 64-bit systems, we use the GDT only minimally. However, the CPU still requires a correctly configured GDT to run in protected mode, even if it is not actually used.

This trait specifies base functionality common to both types of descriptor table.

Associated Types

type Entry: Sized

The type of an entry in this descriptor table.

For an IDT, these are interrupt Gates, while for a GDT or LDT, they are segment Descriptors.

Required Methods

fn entry_count(&self) -> usize

Returns the number of Entries in the DTable.

This is used for calculating the limit.

fn load(&'static self)

Load the descriptor table with the appropriate load instruction

Provided Methods

fn get_ptr(&self) -> Pointer<Self>

Get the IDT pointer struct to pass to lidt or lgdt

This expects that the object implementing DTable not contain additional data before or after the actual DTable, if you wish to attach information to a descriptor table besides the array of entries that it consists of, it will be necessary to encose the descriptor table in another struct or enum type.