Module cpu::interrupts

Intterupt handling on x86 machines.

This module provides support for interrupt handling on both x86 and x86_64 as a black box. Code that depends on this can use the same API regardless of system word size.

Modules

idt
Common functionality for the x86 and x86_64 Interrupt Descriptor Table.

pics
Support for the 8259 Programmable Interrupt Controller.

Structs

ExceptionInfo
A description of a CPU exception

Constants

NUM_EXCEPTIONS
Number of interrupt vectors corresponding to CPU exceptions.

Functions

page_fault
Handles page fault exceptions

test
Test interrupt handler for ensuring that the IDT is configured correctly.

timer
Handler for the system timer interrupt

Type Definitions

ErrorCodeHandler
An ISR that handles an error with an error code

InterruptHandler
An ISR that handles a regular interrupt