Module cpu::interrupts::pics

Support for the 8259 Programmable Interrupt Controller.

The 8259 PIC controls all of our IO IRQs. The PIC recieves IRQ requests and feeds them to the CPU in order. An x86 system typically has two PICs, PIC1 and PIC2, each of which provides 8 IRQs. Two PICs provide us with 15 unique IRQs, because one interrupt on the leader PIC is linked to interrupts on the follower PIC.

Mistakes were made, and Intel boneheadedly decided that it was a good idea to map the vector offset of PIC1 to 8, so that it maps to interrupts 8 ... 15 in the IDT. This conflicts with some of the interrupt numbers used by CPU exceptions. Therefore, we must remap the PIC vectors so that PIC1 starts at 32 and PIC2 at 40.

Structs

PIC
A 8259 Programmable Interrupt Controller.

Enums

IRQ
List of IRQs on the x86.

Functions

end_pic_interrupt⚠
If an interrupt is being handled by the PICs, end that interrupt.

initialize⚠
Initialize the system's Programmable Interrupt Controller