Module cpu::segment

Code for using the x86 and x86_64 segmentation hardware.

For more information, refer to the Intel® 64 and IA-32 Architectures Software Developer’s Manual, Vol. 3A, section 3.2, "Using Segments". Some of the documentation present in this module was taken from the Intel manual.

Structs

CodeFlags
The type-specific section of a code segment's flags

DataFlags
The type-specific section of a data segment's flags

Descriptor
A segment descriptor is an entry in an IDT or GDT.

Flags
Segment descriptor bitflags field.

Gdt
Structure representing a Global Descriptor Table

Selector
A segment selector is a 16-bit identifier for a segment.

Enums

SysType
Possible types of for a system segment.

Type
Possible ways to interpret the type bits of a segment selector.

Constants

ACCESSED
If this is a code or data segment and the accessed bit is set, it has been accessed.

AVAILABLE
1 if this segment is available for use by system software

CODE_ACCESSED
0 if this segment hasn't been accessed, 1 if it has

CODE_DATA_ACC
1 if this is a code or data segment that has been accessed

CONFORMING
0 if this segment is non-conforming, 1 if it is conforming

DATA_ACCESSED
0 if this segment hasn't been accessed, 1 if it has

DEFAULT_SIZE
0 if this is a 16-bit segment, 1 if it is a 32-bit segment

DESCR_TYPE
1 if this is a data/code segment, 0 if this is a system segment

DPL
Two bits indicating the descriptor priveliege level

EXECUTE
0 if this segment is not executable, 1 if it is executable

EXEC_ONLY
Whether this segment is execute-only

EXPAND_DOWN
1 if this segment expands down

GDT_SIZE
The number of entries in the GDT

GRANULARITY
0 if the limit of this segment is given in bytes, 1 if it is given in 4092-byte pages

LENGTH
0 if this is a 16- or 32-bit segment, 1 if it is a 64-bit segment

LIMIT
bits 16...19 of the limit

PRESENT
1 if this segment is present.

READ
0 if this segment is read-only, 1 if it is read-write

RPL
Requested Privelege Level (RPL) bits

RPL_RING_0
Set if the RPL is in Ring 0

RPL_RING_1
Set if the RPL is in Ring 1

RPL_RING_2
Set if the RPL is in Ring 2

RPL_RING_3
Set if the RPL is in Ring 3

SEGMENT_TYPE
Four bits that indcate the type of the segment

TI_GDT
If the Table Indicator (TI) is 0, use the GDT

TI_LDT
If the TI is 1, use the LDT

WRITE
0 if this segment is read-only, 1 if it is read-write

Statics

GDT
A Global Descriptor Table (GDT)