Struct cpu::segment::Descriptor

#[repr(C, packed)]

pub struct Descriptor {
    pub base_high: u8,
    pub flags: Flags,
    pub base_mid: u8,
    pub base_low: u16,
    pub limit: u16,
}

A segment descriptor is an entry in an IDT or GDT.

A segment descriptor is a data structure in a GDT or LDT that provides the processor with the size and location of a segment, as well as access control and status information. Segment descriptors are typically created by compilers, linkers, loaders, or the operating system or executive, but not application programs.

Fields
base_high: u8

The last 8 bits of the base address

flags: Flags

The next 16 bits are bitflags

base_mid: u8

The middle 8 bits of the base address

base_low: u16

The first 16 bits of the base address

limit: u16

the segment limit

Methods

impl Descriptor

const fn null() -> Self

Constructs a new null Descriptor

fn new(base: u32, limit: u32) -> Self

Constructs a new Descriptor from a limit and a base address

fn get_limit(&self) -> u32

Extract the limit part from the flags and limit fields.