Struct cpu::Registers

#[repr(C, packed)]

pub struct Registers {
    pub rsi: u64,
    pub rdi: u64,
    pub r11: u64,
    pub r10: u64,
    pub r9: u64,
    pub r8: u64,
    pub rdx: u64,
    pub rcx: u64,
    pub rax: u64,
}

Registers pushed to the stack when handling an interrupt or context switch.

Fields
rsi: u64 rdi: u64 r11: u64 r10: u64 r9: u64 r8: u64 rdx: u64 rcx: u64 rax: u64
Methods

impl Registers

unsafe fn to_array(&self) -> [u64; 9]

Transform this struct into an array of u64s (if you would ever want to do this) TODO: rewrite this to be a convert::Into implementation.

const fn empty() -> Self

Create a new empty set of Registers

unsafe fn push()

Push the caller-saved registers to the stack (such as when handling a context switch or interrupt).

THIS FUNCTION IS NAKED. DO NOT CALL IT NORMALLY.

unsafe fn pop()

Push the caller-saved registers off the stack (such as when handling a context switch or interrupt).

THIS FUNCTION IS NAKED. DO NOT CALL IT NORMALLY.