Module cpu::timer::timestamp

x86 Timestamp register

Functions

get_timestamp
Returns the current timestamp, or an error

is_available
Returns true if timestamps are currently available.

rtdsc⚠
Read the current value of the timestamp counter.

rtdscp⚠
Read the current timestamp, after other instructions have been executed.

wait_get_timestamp
Returns the current timestamp or an error, after other instructions have been executed.